#223: Move tests for distinct attr mapping
